Adherent with usable levitation plates no longer hurt themselves when falling
:cl:
bugfix: Adherent with usable levitation plates no longer hurt themselves when falling
/:cl:

I noticed that Adherent had a partial `handle_fall_special` implementation, but it had two problems with it:
- It would only trigger if you were floating, which prevents falling in the first place
- Even if it did trigger, it doesn't return TRUE, meaning it will show the flavor text and then offer no fall protection

It looks like the code was intended to provide fall protection similar to that of a GAS, so I've marked this as a bugfix instead of a tweak.

I've tested this locally by jumping down the stairwell, and it works as intended.